app/requirements.txt
wheel==0.34.2
python-libnmap==0.7.0
Flask==2.2.5
Flask-FontAwesome==0.1.4
requests
bidict==0.22.0
cachelib==0.9.0
click==8.1.3
gevent
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SocketIO
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-engineio
python-socketio
six==1.11.0
Werkzeug==2.2.3
zipp==3.8.1
